pyproject.toml
[tool.poetry]
name = "classify-medical-research-papers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "classify_medical_research_papers"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.352"
openai = "^1.6.1"
tiktoken = "^0.5.2"
faiss-cpu = "^1.7.4"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
scikit-learn = "^1.3.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
